# openapi.yml
openapi: 3.0.3
info:
  title: Strand Chat API
  version: "1.0"
servers:
  - url: http://localhost:3001/api
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: strand_auth
...
paths:
...
  /conversations:
    get:
      summary: List conversations
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      summary: Create conversation
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
...
  /push/vapid-public-key:
    get:
      summary: Get VAPID public key
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  publicKey:
                    type: string
  /push/subscribe:
    post:
      summary: Register a push subscription
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                endpoint:
                  type: string
                expirationTime:
                  type: integer
                  nullable: true
                keys:
                  type: object
                  properties:
                    p256dh:
                      type: string
                    auth:
                      type: string
      responses:
        "200":
          description: OK
  /push/unsubscribe:
    post:
      summary: Remove a push subscription
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                endpoint:
                  type: string
      responses:
        "200":
          description: OK
...